firmware: app: tasks: Increase startup event group timeout
This was done because to initialize EDC correctly additional delays had
to be added, precisely because of the shared UART interface between both
EDCs. Naturally, the event group event timeout for the other tasks had
to be updated accordingly.

This commit also makes sure all payloads are disabled once the startup
task exits. This way a centralized payload control is handled by the
mission manager task, which makes sure state is persistent.